Add files via upload
Project Name: Market Predictor

Description:
This is a Python application for educational purposes that allows users to perform various tasks, including data analysis, visualization, and reporting. It uses the pandas and matplotlib libraries for data manipulation and plotting. Please note that this project is intended solely for learning and educational purposes and should not be used in production environments.

Installation:
1. Clone the repository.
2. Install the required libraries using 'pip install -r libraries.txt'.

Technologies: Python, pandas, matplotlib

Usage: 
1. Run 'python main.py' to start the application.

Disclaimer:
This project is provided for educational purposes only. It is not intended for use in real-world applications, and the developers and contributors are not responsible for any consequences resulting from its use in production environments.